Added automated poll system, added auto shop system

+++ main.py
@@ -26,7 +26,6 @@ async def on_ready():
     conn.commit()
     conn.close()
     await leaderboard(client)
-
     
 async def tax():
     channel = client.get_channel(698403873374601237)
@@ -91,7 +90,64 @@ async def bank():
     await channel.send(f"{bankping.mention} A 6% bank fee has been collected. *(No more than $200 was taken from your account)*")
     await leaderboard(client)
 
-            
+async def shopchange():
+    channel = client.get_channel(710211979360338020)
+    conn = sqlite3.connect('hierarchy.db')
+    c = conn.cursor()
+    c.execute('SELECT * FROM shop')
+    stats = c.fetchall()
+    embed = discord.Embed(color=0x30ff56)
+    embed.set_author(name='Shop')
+    x = 1
+    text = ""
+    for stat in stats:
+        change = ""
+        if stat[2] == 'down':
+            if random.randint(1,3) == 1:
+                change = 'up'
+            else:
+                change = 'down'
+        if stat[2] == 'up':
+            if random.randint(1,3) == 1:
+                change = 'down'
+            else:
+                change = 'up'
+        
+        if change is 'up':
+            newmax = stat[1] + 5
+            if newmax > stat[4]:
+                newmax = stat[4]
+            newprice = random.randint(stat[1], newmax)
+      
+        if change is 'down':
+            newmin = stat[1] - 5
+            if newmin < stat[3]:
+                newmin = stat[3]
+            newprice = random.randint(newmin, stat[1])
+        
+
+        if newprice > stat[1]:
+            text = f'{text}\n{stat[0].capitalize()}: ⏫ ${newprice-stat[1]}'
+        elif newprice < stat[1]:
+            text = f'{text}\n{stat[0].capitalize()}: ⏬ ${stat[1]-newprice}'
+        if newprice == stat[1]:
+            text = f'{text}\n{stat[0].capitalize()}: No change.'
+
+
+        c.execute(f"UPDATE shop SET price = {newprice} WHERE name = '{stat[0]}'")
+        c.execute(f"UPDATE shop SET last = '{change}' WHERE name = '{stat[0]}'")
+
+        embed.add_field(name=f'{x}. ${newprice} - {stat[0].capitalize()} {stat[6]}', value=f'{stat[5]}', inline=False)
+        x += 1
+
+
+    await channel.send(text)
+    shopchannel = client.get_channel(702654620291563600)
+    message = await shopchannel.fetch_message(702906494022058084)
+    await message.edit(embed=embed)
+    conn.commit()
+    conn.close()
+
 @tasks.loop(seconds=1)
 async def heisttimer():
     heist = open_json()
@@ -175,9 +231,11 @@ async def heisttimer():
 @tasks.loop(seconds=60)
 async def eventtimer():
     embed = discord.Embed(color=0x442391)
-    channel = client.get_channel(698322322834063430)
+    feechannel = client.get_channel(698322322834063430)
     pollchannel = client.get_channel(698009727803719757)
-    feemessage = await channel.fetch_message(698775208663973940)
+    shopchannel = client.get_channel(710211730797756477)
+    feemessage = await feechannel.fetch_message(698775208663973940)
+    shopmessage = await shopchannel.fetch_message(710213879707336805)
     embed.set_author(name="Fee collection times")
     taxtime = time.localtime()
     minutes = taxtime[4] + taxtime[3]*60
@@ -185,40 +243,71 @@ async def eventtimer():
     banktime = time.localtime()
     minutes = banktime[4] + banktime[3]*60
     banktime = 720-minutes
+    shoptime = time.localtime()
+    minutes = shoptime[4] + shoptime[3]*60
+    times = [180,360,540,720,900,1080,1260]
+    for x in times:
+        if minutes > x and minutes < x+180:
+            shoptime = x+180
+    if minutes in times:
+        minutes = 0
+    shoptime -= minutes
     if banktime < 0 or banktime == 720:
         banktime = 1440-minutes
         if banktime == 1440:
             banktime = 0
     if taxtime == 1440:
         await tax()
-        embed.add_field(name="Tax collection",value=f'{minisplittime(taxtime)}',inline=False)
-    else:
-        embed.add_field(name="Tax collection",value=f'{minisplittime(taxtime)}',inline=False)
+    
+    embed.add_field(name="Tax collection",value=f'{minisplittime(taxtime)}',inline=False)
+    
     if banktime == 0:
         await bank()
         embed.add_field(name="Bank fee collection",value='12h 0m',inline=False)
     else:
         embed.add_field(name="Bank fee collection",value=f'{minisplittime(banktime)}',inline=False)
 
+    embed2 = discord.Embed(color=0x30ff56)
+    embed2.set_author(name='Shop change times')
+    
+    
+    if shoptime == 0:
+        await shopchange()
+        embed2.add_field(name='__________', value='3h 0m', inline=False)
+    else:
+        embed2.add_field(name='__________', value=f'{minisplittime(shoptime)}', inline=False)
             
     await feemessage.edit(embed=embed)
+    await shopmessage.edit(embed=embed2)
     conn = sqlite3.connect('hierarchy.db')
     c = conn.cursor()
     c.execute('SELECT * FROM polls')
     polls = c.fetchall()
     for poll in polls:
-        if poll[1] < time.time():
+        message = await pollchannel.fetch_message(poll[2])    
+        text = message.content
+        text = text[::-1]
+        if poll[1] < time.time() and poll[1] != 0:
             c.execute(f"DELETE FROM polls WHERE name = '{poll[0]}'")
-            conn.commit()
-        else:
-            message = await pollchannel.fetch_message(poll[2])
-            text = message.content
             index = text.index('\n')
             text = text[index:]
             text = text[::-1]
-            text = f"{text}**Time left:{minisplittime(int(poll[1]-int(time.time()))/60)}**"
+            results = ""
+            for reaction in message.reactions:
+                temp = f"{str(reaction.emoji)}: {reaction.count-1}   "
+                results = f"{results}{temp}"
+            text = f'{text}**Poll closed. Results:**\n\n{results}'
             await message.edit(content=text)
+            await message.clear_reactions()
 
+
+        elif poll[1] > time.time() and poll[1] != 0:
+            index = text.index('\n')
+            text = text[index:]
+            text = text[::-1]
+            text = f"{text}**Time left: {minisplittime(int(int(poll[1]-time.time())/60))}**"
+            await message.edit(content=text)
+    conn.commit()
     conn.close()
     
 @client.event
@@ -227,6 +316,7 @@ async def on_raw_reaction_add(payload):
     taxping = guild.get_role(698321954742075504)
     bankping = guild.get_role(698322063206776972)
     tokendm = guild.get_role(706589874966364191)
+    pollchannel = client.get_channel(698009727803719757)
     if payload.message_id == 698774871949181039:
         user = payload.user_id
         user = guild.get_member(user)
@@ -239,6 +329,12 @@ async def on_raw_reaction_add(payload):
         user = payload.user_id
         user = guild.get_member(user)
         await user.add_roles(tokendm)
+    if payload.channel_id == 698009727803719757:
+        if payload.user_id != 698771271353237575:
+            message = await pollchannel.fetch_message(payload.message_id)
+            for reaction in message.reactions:
+                if str(reaction.emoji) != str(payload.emoji):
+                    await client.http.remove_reaction(payload.channel_id, payload.message_id, reaction.emoji, payload.user_id)
 
 @client.event
 async def on_raw_reaction_remove(payload):
